Feat : Create Room 관련 리팩토링
